# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Node first docs"
  },
  "paths": {
    "/users/{userID}": {
      "get": {
        "description": "Get user by ID",
        "tags": [
          "User"
        ],
        "operationId": "getUserById",
        "parameters": [
          {
            "name": "userID",
            "in": "path",
            "type": "string",
            "required": true
          },
          {
            "name": "Authorization",
            "in": "header",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful responce.",
            "schema": {
              "$ref": "#/definitions/SingleUser"
            }
          },
          "400": {
            "description": "Bad request.",
            "schema": {
              "$ref": "#/definitions/ApiError"
            }
          }
        }
      },
      "put": {
        "description": "Update user",
        "tags": [
          "User"
        ],
        "operationId": "UpdateUserById",
        "parameters": [
          {
            "name": "email",
            "in": "formData",
            "type": "string",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "User was updated.",
            "schema": {
              "$ref": "#/definitions/SingleUser"
            }
          },
          "400": {
            "description": "Bad request.",
            "schema": {
              "$ref": "#/definitions/ApiError"
            }
          }
        }
      }
    },
    "/users": {
      "get": {
        "description": "Get all users.",
        "tags": [
          "User"
        ],
        "operationId": "GetAllUsers",
        "responses": {
          "200": {
            "description": "Success responce.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/SingleUser"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "ApiError": {
      "properties": {
        "statusCode": {
          "type": "number",
          "example": 400
        },
        "message": {
          "type": "string",
          "example": "Something went wrong."
        }
      }
    },
    "SingleUser": {
      "properties": {
        "schema": {
          "properties": {
            "name": {
              "type": "string",
              "example": "Viktor"
            },
            "age": {
              "type": "integer",
              "example": 10
            },
            "email": {
              "type": "string",
              "example": "max@gmail.com"
            }
          }
        }
      }
    }
  }
}